Major updates on core resources (ItrackerResources) and language-admin section. Minor fixes on JSPs, internationalizations, permissions. Tests re-enabled or changed.
Fixes some charset issues when using UTF-8 database, databases should be updated manually to support UTF-8.
Language-keys will not be initialized on database, the defaults are consumed from the properties (fallback) and are overridden by database if re-defined in language-admin. (unchanged keys should be removed from database manually by querying on updated date)

+++ src/test/java/org/itracker/persistence/dao/PermissionDAOImplTest.java
@@ -50,7 +50,7 @@ public class PermissionDAOImplTest extends AbstractDependencyInjectionTest {
 		assertNotNull(foundPermission);
 		assertNotNull(foundUser);
 		assertNotNull(foundProject);
-		assertEquals(2, foundPermission.getPermissionType());
+		assertEquals(2, foundPermission.getPermissionType().intValue());
 		assertEquals("admin_test1", foundUser.getLogin());
 		assertEquals("test_name", foundProject.getName());
